Add fixes for firebase authentication
