+++ app.json
@@ -15,6 +15,8 @@
       "supportsTablet": false
     },
     "android": {
+      "package": "com.priyesh.pokedex",
+      "versionCode": 1,
       "adaptiveIcon": {
         "foregroundImage": "./assets/Logo_Pokedex.png",
         "backgroundColor": "#ffffff"
