+++ package.json
@@ -14,11 +14,11 @@
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/native-stack": "^6.9.17",
     "expo": "~49.0.15",
-    "expo-firebase": "^0.0.0-alpha.0",
     "expo-font": "~11.4.0",
     "expo-status-bar": "~1.6.0",
+    "firebase": "^10.8.1",
     "react": "18.2.0",
-    "react-native": "0.72.6",
+    "react-native": "0.72.10",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0"
   },
