--- services/firebaseDatabase.js
@@ -1,3 +1,0 @@
-export const saveUserData = (id, username, password) => {
-    console.log("Need to handle the info !!!", id, username, password);
-};
